Initial version of the app
This brings together all of the techniques that we explored as part of this project. The resulting app loads the astronaut and mission data, presents it in navigable views, and looks at techniques for modelling data (structs and embedded structs with references between data sets) and using generics to make the load process more graceful.