Project 8 challenge 3
Added a navigation bar button to switch between launch date and crew members as secondary information in the mission list. This uses a Mission extension that returns a list of crew IDs which is then used to extract full crew names from the astronaut collection. Also tidied up the astronaut view to add a section header to the mission list.